updating devices with with VLAN profile to be deleted before deletion, no warning popup yet

+++ public/javascripts/showvlanprofiles_actions.js
@@ -124,19 +124,55 @@ $(document).ready(function() {
     cross.removeClass('fa-times').addClass('fa-plus');
   });
 
-  $(document).on('click', '.btn-trash', function(event) {
-    $.ajax({
-      type: 'DELETE',
-      url: '/vlan/profile/del',
-      traditional: true,
-      data: {ids: selectedItens},
-      success: function(res) {
+  $(document).on('click', '.btn-trash', async function(event) {
+    let updatesFailed = false;
+    let deviceFailed;
+    for (let i = 0; i < selectedItens.length; i++) {
+      let res = await $.get('/vlan/profile/check/'+selectedItens[i], 'json');
+      if (res.type == 'success') {
+        res.updateDevices.every((updateObj) => {
+          updateObj = JSON.parse(updateObj);
+          $.ajax({
+            type: 'POST',
+            url: '/vlan/update/'+updateObj.deviceId,
+            traditional: true,
+            data: {
+              vlans: updateObj.vlans,
+            },
+            success: function(res) {
+              return true;
+            },
+            error: function(res) {
+              updatesFailed = true;
+              deviceFailed = updateObj.deviceId;
+              return false;
+            },
+          });
+        });
+      } else {
         displayAlertMsg(res);
-        if (res.type == 'success') {
-          fetchVlanProfiles(vlanProfilesTable);
-        }
-      },
-    });
+      }
+      if (updatesFailed === true) {
+        break;
+      }
+    }
+    if (updatesFailed === true) {
+      displayAlertMsg('Deleção de perfis não foi possível pois dispositivo ' +
+        deviceFailed+' não atualizou.');
+    } else {
+      $.ajax({
+        type: 'DELETE',
+        url: '/vlan/profile/del',
+        traditional: true,
+        data: {ids: selectedItens},
+        success: function(res) {
+          displayAlertMsg(res);
+          if (res.type == 'success') {
+            fetchVlanProfiles(vlanProfilesTable);
+          }
+        },
+      });
+    }
   });
 
   // Use this format when adding button with AJAX
